Move template functions to blueprint
