+++ mfo/app.py
@@ -36,45 +36,14 @@ def create_app():
     import mfo.admin.views
     import mfo.account.views
     import mfo.database.commands
+    import mfo.template_functions
     app.register_blueprint(mfo.home.views.bp)
     app.register_blueprint(mfo.admin.views.bp)
     app.register_blueprint(mfo.account.views.bp)
     app.register_blueprint(mfo.database.commands.bp)
+    app.register_blueprint(mfo.template_functions.bp)
 
-    # Define helper functions
-    @app.template_global()
-    def update_sort(column, sort_by, sort_order):
-        if column in sort_by:
-            # Move the existing column to the end of the list
-            sort_by.remove(column)
-            sort_by.append(column)
-        else:
-            # Or, add the new column to the end of the list
-            sort_by.append(column)
-        return sort_by
-
-    @app.template_global()
-    def update_order(column, sort_by, sort_order):
-        if column in sort_by:
-            # If column altready exists in sort_by list,
-            # find its corresponding sort_order value,
-            # reverse it, and move it to the end of the list
-            index = sort_by.index(column)
-            order = sort_order.pop(index)
-            if order == 'asc':
-                sort_order.append('desc')
-            else:
-                sort_order.append('asc')
-        else:
-            # Or, add the sort_order value for the new column 
-            # to the end of the list
-            sort_order.append('asc')
-        return sort_order
-
-    # Add helper functions to Jinja2 environment
-    # app.jinja_env.globals.update(update_sort=update_sort)
-    # app.jinja_env.globals.update(update_order=update_order)
-
+    
     return app
